Isolate sync ack header checking to HttpCallbacksEventPost class only
