+++ src/test/resources/cloudfwd.properties
@@ -14,9 +14,10 @@
 
 ###### HEC Properties ######
 #HEC url with port (Multiple url's can be comma separated list)
-url=https://inputs1.kinesis4.splunkcloud.com:8088
+# -Dtest=MultiThreadedVolumeTest -DargLine="-Durl=https://127.0.0.1:8088 -Dsplunk_hec_token=ssergeev-nile -Dnum_senders=1 -Dthreads_per_connection=1"
+url=https://127.0.0.1:8088
 #HEC token
-splunk_hec_token=3664A2C4-50B4-485E-86E0-C65D5A4B9232
+splunk_hec_token=ssergeev-nile
 #Override hostname of machine sending data
 splunk_hec_host=
 # Splunk index in which the data feed is stored (Default: main)
@@ -31,7 +32,7 @@ splunk_hec_sourcetype=
 #Mock Splunk flag for test execution
 mock_http=false
 disable_certificate_validation=true
-channels_per_dest=1
+channels_per_dest=2
 max_total_channels=99999999
 max_unacked_per_channel=13
 event_batch_size=4000000
@@ -78,4 +79,4 @@ ssl_cert_content=
 ssl_cert_hostname_regex=
 
 #3 min preflight timeout
-preflight_timeout=60000
+preflight_timeout=600000
